Moved class files into proper directory. createURL and makeConnection methods are ready for testing.
