+++ README.md
@@ -1,3 +1,3 @@
 Project for Android Basics Lesson 4: Networking
 
-Rubric can be found here: https://github.com/seth-wat/Android-Book-Listing.git
+Rubric can be found here: https://github.com/seth-wat/Android-com.example.android.booklisting.Book-Listing.git
